safari-600.7-branch/LayoutTests/fast/css/readwrite-pseudoclass-input-expected.txt

Tests read-write pseudo-class properties are not applied to disabled input and textarea elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.getElementById('inputdisabled'),null).getPropertyValue('background-color') is not "rgb(255, 0, 0)"
PASS window.getComputedStyle(document.getElementById('textareadisabled'),null).getPropertyValue('background-color') is not "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
         
